Revert bug fix 1279577
https://review.openstack.org/73108
The original change does not work for the case where the api-cfn recipe
is included from within another recipe.
This revert does not cause any harm, it justs registers
api-cfn with identity unnecessarily.
The real solution is probably a refactoring of the common identity recipe.
Created blueprint identity-registration-cleanup for this future work.

Change-Id: I43646182a7465cdddd55b1f840764ba1b147451e
Closes-Bug: #1309123