Run heat-manage as correct user/group
We need to explicitly run heat-manage with heat user/group.
Running as root may cause files in /var/log/heat/* to be owned by
root, and this will cause heat services from being able to start (due
to being unable to write to files in that directory).
This is a non-issue when you install from packages as the package
post-install typically does a heat-manage with the correct user, and
this ensures logs are created w/ correct ownership.

Change-Id: I6aede4f2b768cb191527ec27b0c8132c0582ebf2