Updating Berks to 2.0.18
This is to resolve the broken berks config with the supermarket push.
This also fixes the strainerfile for the issue with the location of the
gems and running bundle exec.
Removed the spec_helper code for strainer to work, and updated
chefspec.

Also ran rubocop on this, there were a touch of changes.

Change-Id: I91e9259eb2a34ed4fa01935d7cf75c0946ad9cba