Create role and domain setup for heat template defined users
To setup heat for template defined users, several identity registration
steps need to happen.  The first is creating the role for it.
This can be done with the current identity registration provider.
The next three steps (doc in the code as todo) deal with the domain and
requires keystone v3 (bug-1267137). These are handled by a script shipped
with heat, heat-keystone-setup-domain.

Change-Id: I74154ccae609526ec92b916bac4ff8bc60aed170
Related-Bug: #1369654